requirements.txt
streamlit
simpy
pandas
numpy
plotly
qrcode[pil]
python-docx
